pyproject.toml
[tool.poetry]
name = "splatnlp"
version = "0.1.0"
description = ""
authors = ["Cesar E Garza <cesar@cegarza.com>"]
readme = "README.md"
packages = [
    { include = "splatnlp", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.5.1"
numpy = "<1.29.0"
pandas = "^2.2.2"
tqdm = "^4.66.4"
requests = "^2.32.3"
gensim = "^4.3.2"
orjson = "^3.10.6"
sqlalchemy = "^2.0.31"
psutil = "^6.0.0"
xxhash = "^3.4.1"
pyarrow = "^16.1.0"
scipy = "1.12"
pillow = "^10.4.0"
plotly = "^5.22.0"
nbformat = "^5.10.4"
umap-learn = "^0.5.6"
torch = "^2.3.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 80
